pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
python-dotenv = "^0.14.0"
python-dateutil = "2.7.3"
requests = "2.22.0"
gunicorn = "20.0.4"
pylint = "^2.7.2"
pymongo = {extras = ["srv"], version = "^3.11.4"}
mongomock = "^3.22.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
selenium = "^3.141.0"
selenium-requests = "^1.3"
mock = "^4.0.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
